src/requirements/jetson_orin.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations[imgaug]==1.0.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grad-cam==1.5.4
grpcio==1.62.1
huggingface-hub==0.23.2
idna==3.7
imageio==2.36.0
imgaug==0.4.0
iopath==0.1.10
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightning==2.2.5
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==2.2.2
parso==0.8.4
#pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.3.1
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.3
torch==2.2.2
torch-tb-profiler==0.4.3
torchaudio==2.2.2
torchinfo==1.8.0
torchmetrics==1.4.0.post0
torchvision==0.17.2
tqdm==4.66.2
#traitlets==5.14.3
#triton==2.2.0
ttach==0.0.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
Werkzeug==3.0.2
yacs==0.1.8
yarl==1.9.4

